# access_control_service/access-control-service-openapi.yaml
openapi: 3.0.3
info:
  title: Compliance & Access Control Service API
  description: |
    Микросервис для управления геозонами, правилами доступа и контроля соблюдения политик безопасности.
    Обеспечивает привязку меток к сущностям (сотрудники/оборудование), проверку разрешенных зон и генерацию оповещений о нарушениях.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: security@rtls.example.com
servers:
  - url: http://localhost:8084/api/v1
    description: Локальный сервер разработки
  - url: https://compliance.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /compliance/check:
    post:
      tags:
        - Compliance
      summary: Проверка соблюдения правил для позиции
      description: Проверяет, соответствует ли текущая позиция сущности установленным правилам.
      operationId: checkCompliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceCheckRequest'
            examples:
              singleCheck:
                summary: Проверка одной позиции
                value:
                  entity_id: "emp-001"
                  position:
                    x: 12.5
                    y: 7.8
                    z: 1.2
                    timestamp: "2023-11-28T14:30:00Z"
      responses:
        '200':
          description: Результат проверки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceCheckResult'
  /compliance/check/batch:
    post:
      tags:
        - Compliance
      summary: Массовая проверка соблюдения правил
      description: Проверяет соблюдение правил для массива позиций.
      operationId: checkComplianceBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchComplianceCheckRequest'
      responses:
        '200':
          description: Результаты массовой проверки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchComplianceCheckResult'
  /compliance/violations:
    get:
      tags:
        - Compliance
      summary: Получение истории нарушений
      description: Возвращает историю зафиксированных нарушений правил.
      operationId: getViolations
      parameters:
        - name: start_time
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Начало периода
        - name: end_time
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Конец периода
        - name: entity_id
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по сущности
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Фильтр по серьезности
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Violation'
  /compliance/violations/realtime:
    get:
      tags:
        - Compliance
      summary: WebSocket для получения нарушений в реальном времени
      description: Устанавливает WebSocket соединение для получения уведомлений о нарушениях в реальном времени.
      operationId: getRealtimeViolations
      responses:
        '101':
          description: Switching Protocols
        '500':
          description: Ошибка сервера.
components:
  schemas:
...
    ComplianceCheckRequest:
      type: object
      required:
        - entity_id
        - position
      properties:
        entity_id:
          type: string
          description: ID сущности
        position:
          type: object
          required:
            - x
            - y
            - timestamp
          properties:
            x:
              type: number
              format: float
            y:
              type: number
              format: float
            z:
              type: number
              format: float
              default: 0.0
            timestamp:
              type: string
              format: date-time
    ComplianceCheckResult:
      type: object
      properties:
        entity_id:
          type: string
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            z:
              type: number
            timestamp:
              type: string
              format: date-time
        is_compliant:
          type: boolean
          description: Соответствует ли позиция всем правилам
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          description: Предупреждения (низкая серьезность)
    BatchComplianceCheckRequest:
      type: object
      properties:
        checks:
          type: array
          items:
            $ref: '#/components/schemas/ComplianceCheckRequest'
    BatchComplianceCheckResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ComplianceCheckResult'
        summary:
          type: object
          properties:
            total_checks:
              type: integer
            compliant:
              type: integer
            violations:
              type: integer
            warnings:
              type: integer
    Violation:
      type: object
      required:
        - violation_id
        - rule_id
        - entity_id
        - severity
        - timestamp
      properties:
        violation_id:
          type: string
          format: uuid
        rule_id:
          type: string
          format: uuid
        rule_name:
          type: string
        entity_id:
          type: string
        entity_name:
          type: string
        entity_type:
          type: string
        geofence_id:
          type: string
          format: uuid
        geofence_name:
          type: string
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            z:
              type: number
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
          description: Описание нарушения
        timestamp:
          type: string
          format: date-time
        acknowledged:
          type: boolean
          default: false
          description: Подтверждено ли нарушение оператором
        acknowledged_by:
          type: string
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true

    # --- Модели ошибок ---
...